Refactor: Notification as a root component
